+++ app/src/main/java/com/smashing/app/data/mapper/my/MyMapper.kt
@@ -1,10 +1,19 @@
 package com.smashing.app.data.mapper.my
 
+import com.smashing.app.data.model.addsports.AddSportsInfo
 import com.smashing.app.data.model.my.ActiveUserProfile
 import com.smashing.app.data.model.my.UserProfile
 import com.smashing.app.data.model.my.UserProfileItem
+import com.smashing.app.data.model.profile.MyPageInfo
+import com.smashing.app.data.model.profile.ProfileInfo
+import com.smashing.app.data.model.profile.SportProfile
+import com.smashing.app.data.remote.dto.my.AddSportProfileRequest
+import com.smashing.app.data.remote.dto.my.AllProfileDto
 import com.smashing.app.data.remote.dto.my.GetMyTierProfileResponse
+import com.smashing.app.data.remote.dto.my.MyPageData
+import com.smashing.app.data.type.GenderType
 import com.smashing.app.data.type.SportType
+import com.smashing.app.data.type.SportType.Companion.findSportTypeToSportCode
 import com.smashing.app.data.type.TierType
 
 fun GetMyTierProfileResponse.toUserProfile(): UserProfile =
@@ -32,3 +41,49 @@ private fun GetMyTierProfileResponse.ProfileItemResponse.toUserProfileItem(): Us
         sportCode = SportType.findSportTypeToSportCode(sportCode),
         isActive = isActive,
     )
+
+fun MyPageData.toProfileInfo(): ProfileInfo {
+    val active = this.activeProfile
+
+    return ProfileInfo(
+        profileId = active.profileId,
+        sportType = findSportTypeToSportCode(active.sportCode),
+        tierType = TierType.findTierType(active.tierCode),
+        lp = active.lp,
+        minLp = active.minLp,
+        maxLp = active.maxLp,
+        winCount = active.wins,
+        loseCount = active.losses,
+        reviewCount = active.reviews,
+
+        nickname = this.nickname,
+        genderType = GenderType.findByName(this.gender)
+    )
+}
+
+fun AllProfileDto.toSportProfile(): SportProfile {
+    return SportProfile(
+        profileId = this.profileId,
+        sportType = SportType.findSportTypeToSportCode(this.sportCode),
+        isActive = this.isActive
+    )
+}
+
+fun MyPageData.toMyPageInfo(): MyPageInfo {
+    return MyPageInfo(
+        profileInfo = this.toProfileInfo(),
+        sportProfiles = this.allProfiles.map { it.toSportProfile() }
+    )
+}
+
+fun AddSportsInfo.toRequest():
+        AddSportProfileRequest {
+    val sportCode = requireNotNull(this.selectedSports?.code) { "sportCode is required" }
+    return AddSportProfileRequest(
+        sportCode = sportCode,
+        experienceRange = this.selectedSkill?.skillCode ?: "LT_3_MONTHS",
+    )
+}
+
+
+
