+++ app/src/main/java/com/smashing/app/data/mapper/review/toGameReviewList.kt
@@ -5,6 +5,8 @@ import com.smashing.app.data.model.cursor.Cursor
 import com.smashing.app.data.model.cursor.CursorPage
 import com.smashing.app.data.model.review.GameReview
 import com.smashing.app.data.remote.dto.cursor.CursorDto
+import com.smashing.app.data.remote.dto.review.MyProfileReviewListData
+import com.smashing.app.data.remote.dto.review.MyReviewsDto
 import com.smashing.app.data.remote.dto.review.GetUserRecentReviewListResponse
 
 fun CursorDto<GetUserRecentReviewListResponse>.toGameReviewList(): CursorPage<GameReview> {
@@ -24,5 +26,25 @@ private fun GetUserRecentReviewListResponse.toGameReview(): GameReview {
         opponentNickname = this.opponentNickname,
         createdAt = convertLocalDateTimeToTime(this.createdAt),
         content = this.content,
+    )
+}
+
+fun MyProfileReviewListData.toGameReviewPage(): CursorPage<GameReview> {
+    return CursorPage(
+        items = this.results.map { it.toGameReview() },
+        cursor = Cursor(
+            snapshotAt = this.snapshotAt,
+            nextCursor = this.nextCursor,
+            hasNext = this.hasNext
+        )
+    )
+}
+
+fun MyReviewsDto.toGameReview(): GameReview {
+    return GameReview(
+        gameReviewId = this.gameReviewId,
+        opponentNickname = this.nickname,
+        createdAt = convertLocalDateTimeToTime(this.createdAt),
+        content = this.content
     )
 }
