+++ app/src/main/java/com/smashing/app/data/remote/datasource/api/ReviewRemoteDataSource.kt
@@ -2,6 +2,7 @@ package com.smashing.app.data.remote.datasource.api
 
 import com.smashing.app.data.remote.dto.BaseResponse
 import com.smashing.app.data.remote.dto.cursor.CursorDto
+import com.smashing.app.data.remote.dto.review.MyProfileReviewListData
 import com.smashing.app.data.remote.dto.review.GetUserRecentReviewListResponse
 
 interface ReviewRemoteDataSource {
@@ -11,4 +12,7 @@ interface ReviewRemoteDataSource {
         cursor: String?,
         size: Int?,
     ): BaseResponse<CursorDto<GetUserRecentReviewListResponse>>
+
+    suspend fun getMyGameReviews(cursor: String?, size: Int?): BaseResponse<MyProfileReviewListData>
+
 }
