+++ app/src/main/java/com/smashing/app/data/remote/datasource/impl/ReviewRemoteDataSourceImpl.kt
@@ -3,6 +3,7 @@ package com.smashing.app.data.remote.datasource.impl
 import com.smashing.app.data.remote.datasource.api.ReviewRemoteDataSource
 import com.smashing.app.data.remote.dto.BaseResponse
 import com.smashing.app.data.remote.dto.cursor.CursorDto
+import com.smashing.app.data.remote.dto.review.MyProfileReviewListData
 import com.smashing.app.data.remote.dto.review.GetUserRecentReviewListResponse
 import com.smashing.app.data.remote.service.ReviewService
 import javax.inject.Inject
@@ -23,4 +24,11 @@ class ReviewRemoteDataSourceImpl @Inject constructor(
             cursor = cursor,
             size = size,
         )
+
+    override suspend fun getMyGameReviews(
+        cursor: String?,
+        size: Int?
+    ): BaseResponse<MyProfileReviewListData> {
+        return reviewService.getMyGameReviews(cursor, size)
+    }
 }
