+++ app/src/main/java/com/smashing/app/data/remote/service/ReviewService.kt
@@ -3,6 +3,7 @@ package com.smashing.app.data.remote.service
 import com.smashing.app.data.remote.dto.BaseResponse
 import com.smashing.app.data.remote.dto.cursor.CursorDto
 import com.smashing.app.data.remote.dto.review.GetUserRecentReviewListResponse
+import com.smashing.app.data.remote.dto.review.MyProfileReviewListData
 import retrofit2.http.GET
 import retrofit2.http.Path
 import retrofit2.http.Query
@@ -20,4 +21,11 @@ interface ReviewService {
         @Query("size")
         size: Int?,
     ): BaseResponse<CursorDto<GetUserRecentReviewListResponse>>
+
+    @GET("/api/v1/users/me/reviews/recent")
+    suspend fun getMyGameReviews(
+        @Query("cursor") cursor: String?,
+        @Query("size") size: Int?,
+    ): BaseResponse<MyProfileReviewListData>
+
 }
