+++ app/src/main/java/com/smashing/app/data/repository/impl/MyRepositoryImpl.kt
@@ -1,24 +1,26 @@
 package com.smashing.app.data.repository.impl
 
 import com.smashing.app.core.util.suspendRunCatching
+import com.smashing.app.data.mapper.my.toMyPageInfo
+import com.smashing.app.data.mapper.my.toRequest
 import com.smashing.app.data.mapper.my.toUserProfile
-import com.smashing.app.data.mapper.toGameReviewPage
-import com.smashing.app.data.mapper.toMyPageInfo
-import com.smashing.app.data.mapper.toRequest
 import com.smashing.app.data.model.addsports.AddSportsInfo
-import com.smashing.app.data.model.cursor.CursorPage
 import com.smashing.app.data.model.my.UserProfile
 import com.smashing.app.data.model.profile.MyPageInfo
-import com.smashing.app.data.model.review.GameReview
 import com.smashing.app.data.remote.datasource.api.MyRemoteDataSource
 import com.smashing.app.data.remote.dto.my.MyProfileSwitchRequest
 import com.smashing.app.data.remote.dto.requireData
 import com.smashing.app.data.repository.api.MyRepository
 import javax.inject.Inject
 
-class MyRepositoryImpl @Inject constructor(
-    private val myRemoteDataSource: MyRemoteDataSource,
-) : MyRepository {
+class MyRepositoryImpl : MyRepository {
+
+    private val myRemoteDataSource: MyRemoteDataSource
+
+    @Inject
+    constructor(myRemoteDataSource: MyRemoteDataSource) {
+        this.myRemoteDataSource = myRemoteDataSource
+    }
 
     override suspend fun getMyPageInfo(): Result<MyPageInfo> = suspendRunCatching {
         myRemoteDataSource.getMyProfile()
@@ -26,14 +28,6 @@ class MyRepositoryImpl @Inject constructor(
             .toMyPageInfo()
     }
 
-    override suspend fun getMyGameReviews(
-        cursor: String?,
-        size: Int?
-    ): Result<CursorPage<GameReview>> = suspendRunCatching {
-        myRemoteDataSource.getMyGameReviews(cursor, size)
-            .requireData()
-            .toGameReviewPage()
-    }
 
     override suspend fun switchActiveMyProfile(
         profileId: String
