+++ app/src/main/java/com/smashing/app/data/repository/impl/ReviewRepositoryImpl.kt
@@ -2,6 +2,7 @@ package com.smashing.app.data.repository.impl
 
 import com.smashing.app.core.util.suspendRunCatching
 import com.smashing.app.data.mapper.review.toGameReviewList
+import com.smashing.app.data.mapper.review.toGameReviewPage
 import com.smashing.app.data.model.cursor.CursorPage
 import com.smashing.app.data.model.review.GameReview
 import com.smashing.app.data.remote.datasource.api.ReviewRemoteDataSource
@@ -27,4 +28,13 @@ class ReviewRepositoryImpl @Inject constructor(
                 size = size,
             ).requireData().toGameReviewList()
         }
+    override suspend fun getMyGameReviews(
+        cursor: String?,
+        size: Int?
+    ): Result<CursorPage<GameReview>> = suspendRunCatching {
+        reviewRemoteDataSource.getMyGameReviews(cursor, size)
+            .requireData()
+            .toGameReviewPage()
+    }
+
 }
