+++ app/src/main/java/com/smashing/app/presentation/profile/review/AllReviewViewModel.kt
@@ -105,7 +105,7 @@ class AllReviewViewModel @Inject constructor(
                 nextCursor = null
             }
 
-            myRepository.getMyGameReviews(
+            reviewRepository.getMyGameReviews(
                 cursor = if (isInit) null else nextCursor,
                 size = PAGE_SIZE
             )
